requirements.txt
requests
websocket-client
pynacl
python-dotenv
numpy